requirements.txt
pbkdf2
pycryptodome
pyqt5